+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 
 ## [Unreleased]
 
+### Added
+- New optional Migration params: `commit_author_name` + `commit_author_email`,
+  used to override the git commit author.
+
 
 ## v0.8.0 - 2023-02-24